requirements.txt
Django==5.0
djangorestframework==3.14.0
pandas==2.1.3